clean: HTML要素を生成する関数を分離
